Configure site to download timelapse images
Load cell edit JSON files from a localhost server and download each
day's image automatically after refreshing the heatmap grid.

This isn't something I want to actually deploy, but I wanted to commit
it to a branch so I have it saved if I ever want to make a different
timelapse in the future.